Fixed a skolemization bug.
This bug occurs during skolemization, when a quantifier is skolemized,
but not all occurrences of the quantification variable inside the
quantifier body are replaced with the introduced skolem constant, due
to a weird interplay between caching and node sharing.  This bug is
also non-deterministic, as it depends on the relative order in which
nodes are stored inside a IdentityHashSet (concretely in the
AnnotatedNode.sharedNoded field), which depends on concrete memory
locations of nodes, which can be different at every execution.

In a nutshell, the problem is that the same FreeVariableDetector
instance is used to detect free variables in all shared nodes of an
annotated node.  When there are several shared nodes, which
additionally mutually share some subnodes, after the first visit the
result is going to be cached, so the same result will (incorrectly) be
returned for all subsequent visits.  This is exactly the reason why
this bug depends on the order in which the shared nodes are iterated
over, which is non-deterministic since an IdentityHashSet is used as
an underlying data structure.

The fix simply creates a fresh FreeVariableDetector for each of
the shared nodes.

A regression test (testAleks_03102013) is added to the BugTests unit
test.